requirements.txt
fastapi
uvicorn
requests
httpx
cryptography
beautifulsoup4
huggingface-hub
python-multipart
jax
jaxlib
numpy
elevenlabs
python-dotenv
agno
Pillow
openai
mcp
